[qa] refactoring for sentry
