+++ zou/app/utils/sentry.py
@@ -1,17 +1,16 @@
-import os
-
-try:
-    import sentry_sdk
-    from sentry_sdk.integrations.flask import FlaskIntegration
-    from sentry_sdk.integrations.rq import RqIntegration
-except ModuleNotFoundError:
-    print("sentry_sdk is not installed")
 from zou.app import config
+
+if config.SENTRY_DSN:
+    try:
+        import sentry_sdk
+        from sentry_sdk.integrations.flask import FlaskIntegration
+        from sentry_sdk.integrations.rq import RqIntegration
+    except ModuleNotFoundError:
+        print("Sentry_sdk module not found.")
 
 
 def init_sentry():
     if config.SENTRY_DSN:
-
         sentry_sdk.init(
             dsn=config.SENTRY_DSN,
             integrations=[
